solve ctrl+c bug
if pressing ctrl+c, copy was not working: it removed the copied text from the textarea and replaced it with the first character of the copied text (reason undetermined; ctrl+v was working). 
In the insertALM() function, I added a specification that if c is pressed while ctrl is also pressed, the copy2Clipboard() function should be executed. This has the additional advantage that the hidden characters added in regex mode will be removed before copying to clipboard.