change character separator
In earlier versions, the keyboard first used the unicode character hair space (U+200A) and later the more apt zero-width non-joiner (ZWNJ, U+200C) to separate Arabic characters. From this version, the Mongolian Vowel Separator (U+180E) will be used to separate characters, so that the ZWNJ, which is often used in Persian to separate clitics from other words, remains available for input.
In this version, the ZWNJ character is also added to the keyboard (right of the space bar), as well as the Persian character heh-yeh and the ampersand, caret and dollar signs.